Перенос обработки запроса из виджета в контроллер для соблюдения MVC
