+++ controllers/SiteController.php
@@ -3,7 +3,10 @@
 namespace app\controllers;
 
 use app\models\search\HistorySearch;
+use app\widgets\Export\Export;
 use Yii;
+use yii\helpers\ArrayHelper;
+use yii\helpers\Url;
 use yii\web\Controller;
 
 class SiteController extends Controller
@@ -28,7 +31,14 @@ class SiteController extends Controller
      */
     public function actionIndex()
     {
-        return $this->render('index');
+
+        $modelSearch = new HistorySearch();
+
+        return $this->render('index', [
+            'modelSearch' => $modelSearch,
+            'dataProvider' => $modelSearch->search(Yii::$app->request->queryParams),
+            'linkExport' => $this->getLinkExport(),
+        ]);
     }
 
 
@@ -38,12 +48,26 @@ class SiteController extends Controller
      */
     public function actionExport($exportType)
     {
-        $model = new HistorySearch();
+        $modelSearch = new HistorySearch();
 
         return $this->render('export', [
-            'dataProvider' => $model->search(Yii::$app->request->queryParams),
+            'dataProvider' => $modelSearch->search(Yii::$app->request->queryParams),
+            'model' => $modelSearch,
             'exportType' => $exportType,
-            'model' => $model
         ]);
+    }
+
+    /**
+     * @return string
+     */
+    private function getLinkExport()
+    {
+        $params = Yii::$app->getRequest()->getQueryParams();
+        $params = ArrayHelper::merge([
+            'exportType' => Export::FORMAT_CSV
+        ], $params);
+        $params[0] = 'site/export';
+
+        return Url::to($params);
     }
 }
